fix(libft.h): header

+++ libft.h
@@ -1,22 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_strlen.c                                        :+:      :+:    :+:   */
+/*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: psergio- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/04 19:38:55 by psergio-          #+#    #+#             */
-/*   Updated: 2021/04/05 04:11:20 by psergio-         ###   ########.fr       */
+/*   Updated: 2021/05/17 23:05:31 by psergio-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
 # define LIBFT_H
-
 # include <stddef.h>
 
 size_t	ft_strlen(char *str);
-void	*ft_memset(void *s, int c, long unsigned int n);
+void	*ft_memset(void *s, int c, size_t n);
 void	ft_bzero(void *s, size_t n);
 
 #endif
